+++ src/quo/i_o/__init__.py
@@ -1,3 +1,4 @@
+from quo.i_o.output.widgets import *
 from .termui import (
         confirm,
         echo,