+++ src/quo/widgets/placeholder.py
@@ -1,0 +1,5 @@
+from quo.i_o.output.html import HTML
+
+grey = HTML('<style color="#888888">(Type something)</style>')
+
+cyan =HTML('<style color="#00FFFF">(Type something)</style>')